assignment is completed
